requirements.txt
python-telegram-bot==22.2
Flask==3.1.0
